+++ LaClock/ModSettings/LocaleEN.cs
@@ -31,6 +31,7 @@ namespace LaClock
             }
         }
 
+
         public IEnumerable<KeyValuePair<string, string>> ReadEntries(IList<IDictionaryEntryError> errors, Dictionary<string, int> indexCounts)
         {
             var localeEntries = new Dictionary<string, string>